requirements.txt
dcm2nii
json5
jupyter
nibabel
numpy
opencv-python
Pillow
sacred
scikit-image
SimpleITK
torch # for full alignment 2.2.0
torchvision
torchaudio
packaging
triton
timm
tqdm
pytest 
chardet 
yacs 
termcolor
submitit 
tensorboardX
fvcore
seaborn
